fix(P4): initialize sscratch=0 before enabling interrupts
根因：trap_entry 通过 csrrw sp, sscratch, sp 区分内核态/用户态。
sscratch==0 表示内核态，非 0 表示用户态（值为内核栈地址）。
若 sscratch 未初始化（固件残留值非 0），内核态中断被误判为用户态，
sp 被替换为垃圾值，上下文写坏 → sepc=0x2 → 指令页错误。

修复：
- interrupt_init(): 在 csrw stvec 前执行 csrw sscratch, zero
- interrupt_init_smp(): 从核同样初始化 sscratch=0
- ipi.rs: SBI 错误码改为有符号显示（-6 比 18446744073709551610 可读）

+++ src/arch/riscv64/ipi.rs
@@ -66,7 +66,7 @@ pub fn wake_up_other_cores() {
             log::warn!(
                 "SMP: hart {} 启动失败 (error={}, value={})",
                 hart_id,
-                ret.error,
+                ret.error as isize,
                 ret.value
             );
         }
